express: example -> recipe
