+++ cookbook/recipes/express/patches/.graphqlrc.ts.60300e.patch
@@ -1,0 +1,19 @@
+index 62df77106..44994a7c0 100644
+--- a/templates/skeleton/.graphqlrc.ts
++++ b/templates/skeleton/.graphqlrc.ts
+@@ -17,10 +17,11 @@ export default {
+       ],
+     },
+ 
+-    customer: {
+-      schema: getSchema('customer-account'),
+-      documents: ['./app/graphql/customer-account/*.{ts,tsx,js,jsx}'],
+-    },
++    // Customer account API - commented out for Express recipe
++    // customer: {
++    //   schema: getSchema('customer-account'),
++    //   documents: ['./app/graphql/customer-account/*.{ts,tsx,js,jsx}'],
++    // },
+ 
+     // Add your own GraphQL projects here for CMS, Shopify Admin API, etc.
+   },
