+++ cookbook/recipes/express/patches/entry.server.tsx.c0cb8d.patch
@@ -1,0 +1,117 @@
+index 6f5c4abfc..a407bed64 100644
+--- a/templates/skeleton/app/entry.server.tsx
++++ b/templates/skeleton/app/entry.server.tsx
+@@ -1,53 +1,77 @@
++import {PassThrough} from 'node:stream';
++import type {EntryContext} from 'react-router';
++import {createReadableStreamFromReadable} from '@react-router/node';
+ import {ServerRouter} from 'react-router';
+ import {isbot} from 'isbot';
+-import {renderToReadableStream} from 'react-dom/server';
++import type {RenderToPipeableStreamOptions} from 'react-dom/server';
++import {renderToPipeableStream} from 'react-dom/server';
+ import {
+   createContentSecurityPolicy,
+   type HydrogenRouterContextProvider,
+ } from '@shopify/hydrogen';
+-import type {EntryContext} from 'react-router';
+ 
+-export default async function handleRequest(
++const ABORT_DELAY = 5_000;
++
++export default function handleRequest(
+   request: Request,
+   responseStatusCode: number,
+   responseHeaders: Headers,
+   reactRouterContext: EntryContext,
+   context: HydrogenRouterContextProvider,
+ ) {
+-  const {nonce, header, NonceProvider} = createContentSecurityPolicy({
+-    shop: {
+-      checkoutDomain: context.env.PUBLIC_CHECKOUT_DOMAIN,
+-      storeDomain: context.env.PUBLIC_STORE_DOMAIN,
+-    },
+-  });
+-
+-  const body = await renderToReadableStream(
+-    <NonceProvider>
+-      <ServerRouter
+-        context={reactRouterContext}
+-        url={request.url}
+-        nonce={nonce}
+-      />
+-    </NonceProvider>,
+-    {
+-      nonce,
+-      signal: request.signal,
+-      onError(error) {
+-        console.error(error);
+-        responseStatusCode = 500;
++  return new Promise((resolve, reject) => {
++    const {nonce, header, NonceProvider} = createContentSecurityPolicy({
++      shop: {
++        checkoutDomain: context.env.PUBLIC_CHECKOUT_DOMAIN,
++        storeDomain: context.env.PUBLIC_STORE_DOMAIN,
+       },
+-    },
+-  );
++    });
+ 
+-  if (isbot(request.headers.get('user-agent'))) {
+-    await body.allReady;
+-  }
++    let shellRendered = false;
++    const userAgent = request.headers.get('user-agent');
+ 
+-  responseHeaders.set('Content-Type', 'text/html');
+-  responseHeaders.set('Content-Security-Policy', header);
++    const readyOption: keyof RenderToPipeableStreamOptions =
++      userAgent && isbot(userAgent) ? 'onAllReady' : 'onShellReady';
+ 
+-  return new Response(body, {
+-    headers: responseHeaders,
+-    status: responseStatusCode,
++    const {pipe, abort} = renderToPipeableStream(
++      <NonceProvider>
++        <ServerRouter
++          context={reactRouterContext}
++          url={request.url}
++          nonce={nonce}
++        />
++      </NonceProvider>,
++      {
++        nonce,
++        [readyOption]() {
++          shellRendered = true;
++          const body = new PassThrough();
++          const stream = createReadableStreamFromReadable(body);
++
++          responseHeaders.set('Content-Type', 'text/html');
++          responseHeaders.set('Content-Security-Policy', header);
++
++          resolve(
++            new Response(stream, {
++              headers: responseHeaders,
++              status: responseStatusCode,
++            }),
++          );
++
++          pipe(body);
++        },
++        onShellError(error: unknown) {
++          reject(error);
++        },
++        onError(error: unknown) {
++          responseStatusCode = 500;
++          if (shellRendered) {
++            console.error(error);
++          }
++        },
++      },
++    );
++
++    setTimeout(abort, ABORT_DELAY);
+   });
+-}
++}
+\ No newline at end of file
