+++ cookbook/recipes/express/patches/package.json.cfee98.patch
@@ -1,0 +1,88 @@
+index 0ee1599a1..cdf3f9cac 100644
+--- a/templates/skeleton/package.json
++++ b/templates/skeleton/package.json
+@@ -2,61 +2,52 @@
+   "name": "skeleton",
+   "private": true,
+   "sideEffects": false,
+-  "version": "2025.5.2",
+   "type": "module",
+   "scripts": {
+-    "build": "shopify hydrogen build --codegen",
+-    "dev": "shopify hydrogen dev --codegen",
+-    "preview": "shopify hydrogen preview --build",
+-    "lint": "eslint --no-error-on-unmatched-pattern .",
++    "build": "react-router typegen && react-router build",
++    "dev": "node ./scripts/dev.mjs",
++    "dev:server": "cross-env NODE_ENV=development nodemon --require dotenv/config ./server.mjs --watch ./server.mjs",
++    "start": "cross-env NODE_ENV=production node ./server.mjs",
+     "typecheck": "react-router typegen && tsc --noEmit",
+-    "codegen": "shopify hydrogen codegen && react-router typegen"
++    "codegen": "shopify hydrogen codegen && react-router typegen",
++    "typegen": "react-router typegen"
+   },
+-  "prettier": "@shopify/prettier-config",
+   "dependencies": {
++    "@react-router/express": "7.8.2",
++    "@react-router/node": "7.8.2",
++    "@remix-run/eslint-config": "^2.16.1",
+     "@shopify/hydrogen": "2025.5.0",
++    "compression": "^1.7.4",
++    "cross-env": "^7.0.3",
++    "express": "^4.19.2",
+     "graphql": "^16.10.0",
+     "graphql-tag": "^2.12.6",
+-    "isbot": "^5.1.22",
++    "isbot": "^5.1.21",
++    "morgan": "^1.10.0",
+     "react": "18.3.1",
+     "react-dom": "18.3.1",
+-    "react-router": "7.8.2",
+-    "react-router-dom": "7.8.2"
++    "react-router": "7.8.2"
+   },
+   "devDependencies": {
+-    "@eslint/compat": "^1.2.5",
+-    "@eslint/eslintrc": "^3.2.0",
+     "@eslint/js": "^9.18.0",
+     "@graphql-codegen/cli": "5.0.2",
+     "@react-router/dev": "7.8.2",
+     "@react-router/fs-routes": "7.8.2",
+     "@shopify/cli": "~3.80.4",
+     "@shopify/hydrogen-codegen": "^0.3.3",
+-    "@shopify/mini-oxygen": "^3.2.1",
+-    "@shopify/oxygen-workers-types": "^4.1.6",
+-    "@shopify/prettier-config": "^1.1.2",
+-    "@total-typescript/ts-reset": "^0.6.1",
+-    "@types/eslint": "^9.6.1",
++    "@types/compression": "^1.7.2",
++    "@types/express": "^4.17.17",
++    "@types/morgan": "^1.9.4",
+     "@types/react": "^18.2.22",
+     "@types/react-dom": "^18.2.7",
+-    "@typescript-eslint/eslint-plugin": "^8.21.0",
+-    "@typescript-eslint/parser": "^8.21.0",
+-    "eslint": "^9.18.0",
+-    "eslint-config-prettier": "^10.0.1",
+-    "eslint-import-resolver-typescript": "^3.7.0",
+-    "eslint-plugin-eslint-comments": "^3.2.0",
+-    "eslint-plugin-import": "^2.31.0",
+-    "eslint-plugin-jest": "^28.11.0",
+-    "eslint-plugin-jsx-a11y": "^6.10.2",
+-    "eslint-plugin-react": "^7.37.4",
+-    "eslint-plugin-react-hooks": "^5.1.0",
+-    "globals": "^15.14.0",
+-    "prettier": "^3.4.2",
+-    "typescript": "^5.9.2",
++    "dotenv": "^16.0.3",
++    "nodemon": "^2.0.22",
++    "npm-run-all": "^4.1.5",
++    "typescript": "5.9.2",
+     "vite": "^6.2.4",
+     "vite-tsconfig-paths": "^4.3.1"
+   },
+   "engines": {
+-    "node": ">=18.0.0"
++    "node": ">=20.0.0 <22.0.0"
+   }
+ }
