+++ cookbook/recipes/express/patches/products.$handle.tsx.7d2454.patch
@@ -1,0 +1,264 @@
+index 422a2eb95..061b059ca 100644
+--- a/templates/skeleton/app/routes/products.$handle.tsx
++++ b/templates/skeleton/app/routes/products.$handle.tsx
+@@ -1,50 +1,7 @@
+-import {
+-  redirect,
+-  useLoaderData,
+-} from 'react-router';
++import {useLoaderData} from 'react-router';
+ import type {Route} from './+types/products.$handle';
+-import {
+-  getSelectedProductOptions,
+-  Analytics,
+-  useOptimisticVariant,
+-  getProductOptions,
+-  getAdjacentAndFirstAvailableVariants,
+-  useSelectedOptionInUrlParam,
+-} from '@shopify/hydrogen';
+-import {ProductPrice} from '~/components/ProductPrice';
+-import {ProductImage} from '~/components/ProductImage';
+-import {ProductForm} from '~/components/ProductForm';
+-import {redirectIfHandleIsLocalized} from '~/lib/redirect';
+ 
+-export const meta: Route.MetaFunction = ({data}) => {
+-  return [
+-    {title: `Hydrogen | ${data?.product.title ?? ''}`},
+-    {
+-      rel: 'canonical',
+-      href: `/products/${data?.product.handle}`,
+-    },
+-  ];
+-};
+-
+-export async function loader(args: Route.LoaderArgs) {
+-  // Start fetching non-critical data without blocking time to first byte
+-  const deferredData = loadDeferredData(args);
+-
+-  // Await the critical data required to render initial state of the page
+-  const criticalData = await loadCriticalData(args);
+-
+-  return {...deferredData, ...criticalData};
+-}
+-
+-/**
+- * Load data necessary for rendering content above the fold. This is the critical data
+- * needed to render the page. If it's unavailable, the whole page should 400 or 500 error.
+- */
+-async function loadCriticalData({
+-  context,
+-  params,
+-  request,
+-}: Route.LoaderArgs) {
++export async function loader({params, context}: Route.LoaderArgs) {
+   const {handle} = params;
+   const {storefront} = context;
+ 
+@@ -52,188 +9,42 @@ async function loadCriticalData({
+     throw new Error('Expected product handle to be defined');
+   }
+ 
+-  const [{product}] = await Promise.all([
+-    storefront.query(PRODUCT_QUERY, {
+-      variables: {handle, selectedOptions: getSelectedProductOptions(request)},
+-    }),
+-    // Add other queries here, so that they are loaded in parallel
+-  ]);
++  const {product} = await storefront.query(PRODUCT_QUERY, {
++    variables: {handle},
++  });
+ 
+   if (!product?.id) {
+     throw new Response(null, {status: 404});
+   }
+ 
+-  // The API handle might be localized, so redirect to the localized handle
+-  redirectIfHandleIsLocalized(request, {handle, data: product});
+-
+-  return {
+-    product,
+-  };
+-}
+-
+-/**
+- * Load data for rendering content below the fold. This data is deferred and will be
+- * fetched after the initial page load. If it's unavailable, the page should still 200.
+- * Make sure to not throw any errors here, as it will cause the page to 500.
+- */
+-function loadDeferredData({context, params}: Route.LoaderArgs) {
+-  // Put any API calls that is not critical to be available on first page render
+-  // For example: product reviews, product recommendations, social feeds.
+-
+-  return {};
++  return {product};
+ }
+ 
+ export default function Product() {
+-  const {product} = useLoaderData<typeof loader>();
+-
+-  // Optimistically selects a variant with given available variant information
+-  const selectedVariant = useOptimisticVariant(
+-    product.selectedOrFirstAvailableVariant,
+-    getAdjacentAndFirstAvailableVariants(product),
+-  );
+-
+-  // Sets the search param to the selected variant without navigation
+-  // only when no search params are set in the url
+-  useSelectedOptionInUrlParam(selectedVariant.selectedOptions);
+-
+-  // Get the product options array
+-  const productOptions = getProductOptions({
+-    ...product,
+-    selectedOrFirstAvailableVariant: selectedVariant,
+-  });
+-
+-  const {title, descriptionHtml} = product;
++  const {
++    product: {title, descriptionHtml},
++  } = useLoaderData<typeof loader>();
+ 
+   return (
+     <div className="product">
+-      <ProductImage image={selectedVariant?.image} />
+-      <div className="product-main">
+-        <h1>{title}</h1>
+-        <ProductPrice
+-          price={selectedVariant?.price}
+-          compareAtPrice={selectedVariant?.compareAtPrice}
+-        />
+-        <br />
+-        <ProductForm
+-          productOptions={productOptions}
+-          selectedVariant={selectedVariant}
+-        />
+-        <br />
+-        <br />
+-        <p>
+-          <strong>Description</strong>
+-        </p>
+-        <br />
+-        <div dangerouslySetInnerHTML={{__html: descriptionHtml}} />
+-        <br />
+-      </div>
+-      <Analytics.ProductView
+-        data={{
+-          products: [
+-            {
+-              id: product.id,
+-              title: product.title,
+-              price: selectedVariant?.price.amount || '0',
+-              vendor: product.vendor,
+-              variantId: selectedVariant?.id || '',
+-              variantTitle: selectedVariant?.title || '',
+-              quantity: 1,
+-            },
+-          ],
+-        }}
+-      />
++      <h1>{title}</h1>
++      <br />
++      <p>
++        <strong>Description</strong>
++      </p>
++      <br />
++      <div dangerouslySetInnerHTML={{__html: descriptionHtml}} />
++      <br />
+     </div>
+   );
+ }
+ 
+-const PRODUCT_VARIANT_FRAGMENT = `#graphql
+-  fragment ProductVariant on ProductVariant {
+-    availableForSale
+-    compareAtPrice {
+-      amount
+-      currencyCode
+-    }
+-    id
+-    image {
+-      __typename
+-      id
+-      url
+-      altText
+-      width
+-      height
+-    }
+-    price {
+-      amount
+-      currencyCode
+-    }
+-    product {
+-      title
+-      handle
+-    }
+-    selectedOptions {
+-      name
+-      value
+-    }
+-    sku
+-    title
+-    unitPrice {
+-      amount
+-      currencyCode
+-    }
+-  }
+-` as const;
+-
+-const PRODUCT_FRAGMENT = `#graphql
+-  fragment Product on Product {
+-    id
+-    title
+-    vendor
+-    handle
+-    descriptionHtml
+-    description
+-    encodedVariantExistence
+-    encodedVariantAvailability
+-    options {
+-      name
+-      optionValues {
+-        name
+-        firstSelectableVariant {
+-          ...ProductVariant
+-        }
+-        swatch {
+-          color
+-          image {
+-            previewImage {
+-              url
+-            }
+-          }
+-        }
+-      }
+-    }
+-    selectedOrFirstAvailableVariant(selectedOptions: $selectedOptions, ignoreUnknownOptions: true, caseInsensitiveMatch: true) {
+-      ...ProductVariant
+-    }
+-    adjacentVariants (selectedOptions: $selectedOptions) {
+-      ...ProductVariant
+-    }
+-    seo {
+-      description
+-      title
+-    }
+-  }
+-  ${PRODUCT_VARIANT_FRAGMENT}
+-` as const;
+-
+ const PRODUCT_QUERY = `#graphql
+-  query Product(
+-    $country: CountryCode
+-    $handle: String!
+-    $language: LanguageCode
+-    $selectedOptions: [SelectedOptionInput!]!
+-  ) @inContext(country: $country, language: $language) {
++  query Product( $handle: String!) {
+     product(handle: $handle) {
+-      ...Product
++      id
++      title
++      descriptionHtml
+     }
+   }
+-  ${PRODUCT_FRAGMENT}
+ ` as const;
