+++ cookbook/recipes/express/patches/root.tsx.0ad938.patch
@@ -1,0 +1,194 @@
+index 375fc7b04..4020ecc59 100644
+--- a/templates/skeleton/app/root.tsx
++++ b/templates/skeleton/app/root.tsx
+@@ -11,46 +11,20 @@ import {
+   useRouteLoaderData,
+ } from 'react-router';
+ import type {Route} from './+types/root';
+-import favicon from '~/assets/favicon.svg';
+-import {FOOTER_QUERY, HEADER_QUERY} from '~/lib/fragments';
+-import resetStyles from '~/styles/reset.css?url';
+-import appStyles from '~/styles/app.css?url';
+-import {PageLayout} from './components/PageLayout';
++import styles from './styles/app.css?url';
+ 
+ export type RootLoader = typeof loader;
+ 
+-/**
+- * This is important to avoid re-fetching root queries on sub-navigations
+- */
+ export const shouldRevalidate: ShouldRevalidateFunction = ({
+   formMethod,
+   currentUrl,
+   nextUrl,
+ }) => {
+-  // revalidate when a mutation is performed e.g add to cart, login...
+   if (formMethod && formMethod !== 'GET') return true;
+-
+-  // revalidate when manually revalidating via useRevalidator
+   if (currentUrl.toString() === nextUrl.toString()) return true;
+-
+-  // Defaulting to no revalidation for root loader data to improve performance.
+-  // When using this feature, you risk your UI getting out of sync with your server.
+-  // Use with caution. If you are uncomfortable with this optimization, update the
+-  // line below to `return defaultShouldRevalidate` instead.
+-  // For more details see: https://remix.run/docs/en/main/route/should-revalidate
+   return false;
+ };
+ 
+-/**
+- * The main and reset stylesheets are added in the Layout component
+- * to prevent a bug in development HMR updates.
+- *
+- * This avoids the "failed to execute 'insertBefore' on 'Node'" error
+- * that occurs after editing and navigating to another page.
+- *
+- * It's a temporary fix until the issue is resolved.
+- * https://github.com/remix-run/remix/issues/9242
+- */
+ export function links() {
+   return [
+     {
+@@ -61,18 +35,19 @@ export function links() {
+       rel: 'preconnect',
+       href: 'https://shop.app',
+     },
+-    {rel: 'icon', type: 'image/svg+xml', href: favicon},
++    {rel: 'icon', type: 'image/svg+xml', href: '/favicon.svg'},
+   ];
+ }
+ 
+-export async function loader(args: Route.LoaderArgs) {
+-  // Start fetching non-critical data without blocking time to first byte
+-  const deferredData = loadDeferredData(args);
++export async function loader({context}: Route.LoaderArgs) {
++  const [customerAccessToken, cartId] = await Promise.all([
++    context.session.get('customerAccessToken'),
++  ]);
+ 
+-  // Await the critical data required to render initial state of the page
+-  const criticalData = await loadCriticalData(args);
++  const deferredData = loadDeferredData({context});
++  const criticalData = await loadCriticalData({context});
+ 
+-  const {storefront, env} = args.context;
++  const {storefront, env} = context;
+ 
+   return {
+     ...deferredData,
+@@ -86,59 +61,29 @@ export async function loader(args: Route.LoaderArgs) {
+       checkoutDomain: env.PUBLIC_CHECKOUT_DOMAIN,
+       storefrontAccessToken: env.PUBLIC_STOREFRONT_API_TOKEN,
+       withPrivacyBanner: false,
+-      // localize the privacy banner
+-      country: args.context.storefront.i18n.country,
+-      language: args.context.storefront.i18n.language,
++      country: storefront.i18n.country,
++      language: storefront.i18n.language,
+     },
++    isLoggedIn: Boolean(customerAccessToken),
+   };
+ }
+ 
+-/**
+- * Load data necessary for rendering content above the fold. This is the critical data
+- * needed to render the page. If it's unavailable, the whole page should 400 or 500 error.
+- */
+-async function loadCriticalData({context}: Route.LoaderArgs) {
++async function loadCriticalData({context}: Pick<Route.LoaderArgs, 'context'>) {
+   const {storefront} = context;
+ 
+-  const [header] = await Promise.all([
+-    storefront.query(HEADER_QUERY, {
+-      cache: storefront.CacheLong(),
+-      variables: {
+-        headerMenuHandle: 'main-menu', // Adjust to your header menu handle
+-      },
+-    }),
+-    // Add other queries here, so that they are loaded in parallel
+-  ]);
++  const layout = await storefront.query(LAYOUT_QUERY, {
++    cache: storefront.CacheLong(),
++  });
+ 
+-  return {header};
++  return {layout};
+ }
+ 
+-/**
+- * Load data for rendering content below the fold. This data is deferred and will be
+- * fetched after the initial page load. If it's unavailable, the page should still 200.
+- * Make sure to not throw any errors here, as it will cause the page to 500.
+- */
+-function loadDeferredData({context}: Route.LoaderArgs) {
+-  const {storefront, customerAccount, cart} = context;
++function loadDeferredData({
++  context,
++}: Pick<Route.LoaderArgs, 'context'> & {cartId?: string}) {
++  const {storefront, cart} = context;
+ 
+-  // defer the footer query (below the fold)
+-  const footer = storefront
+-    .query(FOOTER_QUERY, {
+-      cache: storefront.CacheLong(),
+-      variables: {
+-        footerMenuHandle: 'footer', // Adjust to your footer menu handle
+-      },
+-    })
+-    .catch((error: Error) => {
+-      // Log query errors, but don't throw them so the page can still render
+-      console.error(error);
+-      return null;
+-    });
+-  return {
+-    cart: cart.get(),
+-    isLoggedIn: customerAccount.isLoggedIn(),
+-    footer,
+-  };
++  return {cart: cart.get()};
+ }
+ 
+ export function Layout({children}: {children?: React.ReactNode}) {
+@@ -150,8 +95,7 @@ export function Layout({children}: {children?: React.ReactNode}) {
+       <head>
+         <meta charSet="utf-8" />
+         <meta name="viewport" content="width=device-width,initial-scale=1" />
+-        <link rel="stylesheet" href={resetStyles}></link>
+-        <link rel="stylesheet" href={appStyles}></link>
++        <link rel="stylesheet" href={styles}></link>
+         <Meta />
+         <Links />
+       </head>
+@@ -162,7 +106,11 @@ export function Layout({children}: {children?: React.ReactNode}) {
+             shop={data.shop}
+             consent={data.consent}
+           >
+-            <PageLayout {...data}>{children}</PageLayout>
++            <div className="PageLayout">
++              <h1>{data.layout?.shop?.name} (Express example)</h1>
++              <h2>{data.layout?.shop?.description}</h2>
++              {children}
++            </div>
+           </Analytics.Provider>
+         ) : (
+           children
+@@ -202,3 +150,19 @@ export function ErrorBoundary() {
+     </div>
+   );
+ }
++
++const CART_QUERY = `#graphql
++  query CartQuery($cartId: ID!, $country: CountryCode, $language: LanguageCode)
++  @inContext(country: $country, language: $language) {
++    cart(id: $cartId) {
++  }
++` as const;
++
++const LAYOUT_QUERY = `#graphql
++  query layout {
++    shop {
++      name
++      description
++    }
++  }
++` as const;
