--- examples/express/.env
@@ -1,4 +1,0 @@
-SESSION_SECRET="foobar"
-PUBLIC_STOREFRONT_API_TOKEN="3b580e70970c4528da70c98e097c2fa0"
-PUBLIC_STORE_DOMAIN="hydrogen-preview.myshopify.com"
-PUBLIC_CHECKOUT_DOMAIN="checkout.hydrogen.shop"
