--- examples/express/.gitignore
@@ -1,30 +1,0 @@
-# React Router generated files
-.react-router/
-build/
-
-# Environment variables
-.env
-.env.local
-.env.production.local
-.env.development.local
-.env.test.local
-
-# Dependencies
-node_modules/
-
-# IDE
-.vscode/
-.idea/
-
-# OS
-.DS_Store
-Thumbs.db
-
-# Logs
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# TypeScript
-*.tsbuildinfo
