--- examples/express/app/entry.server.tsx
@@ -1,147 +1,0 @@
-/**
- * By default, Remix will handle generating the HTTP Response for you.
- * You are free to delete this file if you'd like to, but if you ever want it revealed again, you can run `npx remix reveal` ✨
- * For more information, see https://remix.run/file-conventions/entry.server
- */
-
-import {PassThrough} from 'node:stream';
-import {createReadableStreamFromReadable} from '@react-router/node';
-
-import type {AppLoadContext, EntryContext} from 'react-router';
-import {ServerRouter} from 'react-router';
-import {isbot} from 'isbot';
-import {renderToPipeableStream} from 'react-dom/server';
-import {createContentSecurityPolicy} from '@shopify/hydrogen';
-
-const ABORT_DELAY = 5_000;
-
-export default function handleRequest(
-  request: Request,
-  responseStatusCode: number,
-  responseHeaders: Headers,
-  reactRouterContext: EntryContext,
-  loadContext: AppLoadContext,
-) {
-  return isbot(request.headers.get('user-agent'))
-    ? handleBotRequest(
-        request,
-        responseStatusCode,
-        responseHeaders,
-        reactRouterContext,
-        loadContext,
-      )
-    : handleBrowserRequest(
-        request,
-        responseStatusCode,
-        responseHeaders,
-        reactRouterContext,
-      );
-}
-
-function handleBotRequest(
-  request: Request,
-  responseStatusCode: number,
-  responseHeaders: Headers,
-  reactRouterContext: EntryContext,
-  context: AppLoadContext,
-) {
-  return new Promise((resolve, reject) => {
-    const {nonce, header, NonceProvider} = createContentSecurityPolicy({
-      shop: {
-        checkoutDomain: context.env.PUBLIC_CHECKOUT_DOMAIN,
-        storeDomain: context.env.PUBLIC_STORE_DOMAIN,
-      },
-    });
-
-    const {pipe, abort} = renderToPipeableStream(
-      <NonceProvider>
-        <ServerRouter
-          context={reactRouterContext}
-          url={request.url}
-          nonce={nonce}
-        />
-      </NonceProvider>,
-      {
-        nonce,
-        onAllReady() {
-          const body = new PassThrough();
-          const stream = createReadableStreamFromReadable(body);
-
-          responseHeaders.set('Content-Type', 'text/html');
-          responseHeaders.set('Content-Security-Policy', header);
-
-          resolve(
-            new Response(stream, {
-              headers: responseHeaders,
-              status: responseStatusCode,
-            }),
-          );
-
-          pipe(body);
-        },
-        onShellError(error: unknown) {
-          reject(error);
-        },
-        onError(error: unknown) {
-          responseStatusCode = 500;
-          console.error(
-            (error as Error)?.stack ? (error as Error).stack : error,
-          );
-        },
-      },
-    );
-
-    setTimeout(abort, ABORT_DELAY);
-  });
-}
-
-function handleBrowserRequest(
-  request: Request,
-  responseStatusCode: number,
-  responseHeaders: Headers,
-  reactRouterContext: EntryContext,
-) {
-  const {nonce, header, NonceProvider} = createContentSecurityPolicy();
-
-  return new Promise((resolve, reject) => {
-    const {pipe, abort} = renderToPipeableStream(
-      <NonceProvider>
-        <ServerRouter
-          context={reactRouterContext}
-          url={request.url}
-          nonce={nonce}
-        />
-      </NonceProvider>,
-      {
-        nonce,
-        onShellReady() {
-          const body = new PassThrough();
-          const stream = createReadableStreamFromReadable(body);
-
-          responseHeaders.set('Content-Type', 'text/html');
-          responseHeaders.set('Content-Security-Policy', header);
-
-          resolve(
-            new Response(stream, {
-              headers: responseHeaders,
-              status: responseStatusCode,
-            }),
-          );
-
-          pipe(body);
-        },
-        onShellError(error: unknown) {
-          reject(error);
-        },
-        onError(error: unknown) {
-          console.error(
-            (error as Error)?.stack ? (error as Error).stack : error,
-          );
-          responseStatusCode = 500;
-        },
-      },
-    );
-
-    setTimeout(abort, ABORT_DELAY);
-  });
-}
