--- examples/express/app/routes/_index.tsx
@@ -1,26 +1,0 @@
-import {useRouteError, isRouteErrorResponse, Link} from 'react-router';
-
-export default function Index() {
-  return (
-    <>
-      <p>
-        Edit this route in <em>app/routes/index.tsx</em>.
-      </p>
-      <p>
-        <Link to="/products/v2-snowboard">Snowboard product</Link>
-      </p>
-    </>
-  );
-}
-
-export function ErrorBoundary() {
-  const error = useRouteError();
-
-  if (isRouteErrorResponse(error)) {
-    console.error(error.status, error.statusText, error.data);
-    return <div>Route Error</div>;
-  } else {
-    console.error((error as Error).message);
-    return <div>Thrown Error</div>;
-  }
-}
