--- examples/express/app/styles/app.css
@@ -1,38 +1,0 @@
-body {
-  padding: 0;
-  margin: 0;
-  background: rgb(245, 245, 241);
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-    Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-}
-
-h1,
-h2,
-p {
-  margin: 0;
-  padding: 0;
-}
-
-h1 {
-  font-size: 1.6rem;
-  font-weight: 700;
-  margin-bottom: 1rem;
-  line-height: 1.4;
-}
-
-h2 {
-  font-size: 1.2rem;
-  font-weight: 700;
-  margin-bottom: 1rem;
-  line-height: 1.4;
-}
-
-p {
-  font-size: 1rem;
-  line-height: 1.4;
-}
-
-.Layout {
-  padding: 2rem;
-  max-width: 25rem;
-}
