--- examples/express/eslint.config.js
@@ -1,1 +1,0 @@
-export {default} from '../../templates/skeleton/eslint.config.js';
