--- examples/express/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "example-hydrogen-express",
-  "private": true,
-  "sideEffects": false,
-  "type": "module",
-  "scripts": {
-    "build": "react-router typegen && react-router build",
-    "dev": "cross-env NODE_ENV=development nodemon --require dotenv/config ./server.mjs --watch ./server.mjs",
-    "start": "cross-env NODE_ENV=production node ./server.mjs",
-    "typecheck": "react-router typegen && tsc --noEmit"
-  },
-  "dependencies": {
-    "@react-router/express": "7.8.2",
-    "@react-router/node": "7.8.2",
-    "@remix-run/eslint-config": "^2.16.1",
-    "@shopify/hydrogen": "2025.5.0",
-    "compression": "^1.7.4",
-    "cross-env": "^7.0.3",
-    "express": "^4.19.2",
-    "isbot": "^5.1.21",
-    "morgan": "^1.10.0",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-router": "7.8.2"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.18.0",
-    "@react-router/dev": "7.8.2",
-    "@react-router/fs-routes": "7.8.2",
-    "@shopify/cli": "~3.80.4",
-    "@types/compression": "^1.7.2",
-    "@types/express": "^4.17.17",
-    "@types/morgan": "^1.9.4",
-    "@types/react": "^18.2.22",
-    "@types/react-dom": "^18.2.7",
-    "dotenv": "^16.0.3",
-    "nodemon": "^2.0.22",
-    "npm-run-all": "^4.1.5",
-    "typescript": "5.9.2",
-    "vite": "^6.2.4",
-    "vite-tsconfig-paths": "^4.3.1"
-  },
-  "engines": {
-    "node": ">=20.0.0"
-  }
-}
