--- examples/express/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "include": ["app/**/*.d.ts", "app/**/*.ts", "app/**/*.tsx", "*.ts", "*.d.ts", "env.d.ts"],
-  "exclude": [".react-router/**/*", "build/**/*", "node_modules/**/*", "packages/**/*"],
-  "compilerOptions": {
-    "lib": ["DOM", "DOM.Iterable", "ES2022"],
-    "isolatedModules": true,
-    "esModuleInterop": true,
-    "jsx": "react-jsx",
-    "moduleResolution": "Bundler",
-    "resolveJsonModule": true,
-    "module": "ES2022",
-    "target": "ES2022",
-    "strict": true,
-    "allowJs": true,
-    "forceConsistentCasingInFileNames": true,
-    "skipLibCheck": true,
-    "baseUrl": ".",
-    "paths": {
-      "~/*": ["app/*"]
-    },
-    "noEmit": true
-  }
-}
